refactor gateway audio out to a new component
